requirements.txt
torch
torchvision
transformers
uvicorn
opencv-python
langchain
langchain_core
langchain_community
langchain_huggingface
langchain_text_splitters
gradio
python-docx
pymupdf
rapidocr_onnxruntime
unstructured
pillow
accelerate
tiktoken
faiss-cpu
pycocotools
timm
accelerate
bitsandbytes